feat: Add Hacker News Title for Page 1
In this commit, we have added a feature to display the title "News 1" in the Hacker News section when the page number is equal to 1. This enhancement includes appending a corresponding table row with the title to the 'hackerNewsData' array.

This change ensures that the title "News 1" is correctly displayed when the user is on the first page of the Hacker News section, improving the user experience.